refactor(wgmma): carry the pipeline wait depth as a domain attribute
Replace the bare u32 IntegerAttr on
nvvm.wgmma_mma_pipeline_values_m64n64k16_f32_bf16 with a first-class
WgmmaMaxPendingAttr(u8) defined via #[pliron_attr] and declared on the
op through the attributes = (...) macro slot. The 1..=7 invariant now
lives on the attribute's own Verify impl; the op verifier requires the
attribute and delegates the range check, so N=0 and N>7 are rejected
with a clear message. The u8 depth is threaded through the mir-lower
pipeline plan and template, and ops_test asserts the rejection
messages for 0, 8, and a missing attribute plus the attribute's own
verify bounds.

+++ crates/mir-lower/src/convert/intrinsics/wgmma.rs
@@ -173,7 +173,7 @@ fn pipeline_accumulator_operand_list(slot: usize) -> String {
         .join(", ")
 }
 
-fn pipeline_template(slot_count: usize, group_count: usize, max_pending_groups: u32) -> String {
+fn pipeline_template(slot_count: usize, group_count: usize, max_pending_groups: u8) -> String {
     let result_count = slot_count * VALUE_ACCUMULATOR_COUNT;
     let descriptor_base = result_count * 2;
     let mut template = String::from("{\n    wgmma.fence.sync.aligned;\n");
@@ -404,7 +404,7 @@ pub(crate) fn convert_mma_pipeline_values(
     if !(1..=7).contains(&max_pending_groups) {
         return pliron::input_err_noloc!("pipeline WGMMA max_pending_groups must be in 1..=7");
     }
-    let slot_count = max_pending_groups as usize + 1;
+    let slot_count = usize::from(max_pending_groups) + 1;
     if result_count != slot_count * VALUE_ACCUMULATOR_COUNT {
         return pliron::input_err_noloc!(
             "pipeline WGMMA requires max_pending_groups + 1 accumulator slots"
